Add tests for the location picker view
