+++ packages/apps/esm-login-app/src/root.component.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
 import ChangePassword from './change-password/change-password.component';
-import LocationPicker from './location-picker/location-picker.component';
 import Login from './login/login.component';
+import LocationPickerView from './location-picker/location-picker-view.component';
 import RedirectLogout from './redirect-logout/redirect-logout.component';
 
 export interface RootProps {}
@@ -13,7 +13,7 @@ const Root: React.FC<RootProps> = () => {
       <Routes>
         <Route path="login" element={<Login />} />
         <Route path="login/confirm" element={<Login />} />
-        <Route path="login/location" element={<LocationPicker />} />
+        <Route path="login/location" element={<LocationPickerView />} />
         <Route path="logout" element={<RedirectLogout />} />
         <Route path="change-password" element={<ChangePassword />} />
       </Routes>
